# Conflicts:
#	analyzer/windows/bin/inject-x64.exe
#	analyzer/windows/bin/inject-x86.exe
#	analyzer/windows/bin/monitor-x64.dll
#	analyzer/windows/bin/monitor-x86.dll
#	web/submission/views.py
#	web/templates/submission/status.html

+++ distributed/distributed/views/api.py
@@ -46,6 +46,7 @@ def node_get(name=None):
             enabled=node.enabled,
             name=node.name,
             url=node.url,
+            mode=node.mode,
             machines=machines,
         )
 
@@ -252,8 +253,8 @@ def task_delete(task_id):
         if os.path.isfile(path):
             os.unlink(path)
 
-    # Remove the sample related to this task.
-    if os.path.isfile(task.path):
+    # Remove the sample related to this task (if there's any).
+    if task.path and os.path.isfile(task.path):
         os.unlink(task.path)
 
     # If the task has been finalized then we set the status as deleted. But
